Add wheel support for nv-ingest service (#759)
